Ability to override advertised IP, port and hostname.
In some cases (for example, a service hosted on Mesos), the service
may be running on a different port or even host than the persistent
one where the clients should actually be connecting.

This patch, therefore, allows advertising different port, hostname,
as well as IP address than the listening address.